pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hutwatch"
version = "0.1.0"
description = "BLE temperature monitoring with Telegram bot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bleak>=0.21.0",
    "python-telegram-bot>=20.0",
    "pyyaml>=6.0",
]

[project.scripts]
hutwatch = "hutwatch.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
